🚨 [WIP] 🚨  Make the helper start using the new Version* models so that we slowly transition to their use.
🚨  THIS COMMIT CONTAINS WIP and the code it contains will likely crash at runtime as the changes I started to do on that file were not finished. But since the project got put on hold, I considered it was still worth pushing the WIP in order to be able to pick back up from where I was when it will be time to start working on this again. 🚨

The end goal long term would be to get rid of the helper completely once the models already provide all the necessary helper instance methods to directly do every math and manipulations related to versioning. But in the meantime, keeping the helper as legacy bridge between the two will help making the refactoring easier and smoother while transitioning.