Defaults SwiftUI option to false
After experimenting with it on our current codebase, it seems that this `genstrings` option is quite brittle, with a rather weak RegEx internally not taking semantics into account – which leads to genstrings recognizing code like `Aztec.TextView(…)` as it it was `SwiftUI.Text(…)`… and failing with errors. 🙄

So better turn it off by default after all.