Track result of running `bundle` on M1 chip
Run on my M1 MacMini test device.

Either I didn't run `bundle install` at the time of `f509e58`, or we had
to run it on an M1 first in order for the `nokogiri` platform-agnostic
value to be picked up, or there was some bug in Bundler itself that
prevent that from happening. I'm mentioning the chance of a bug in
Bundler because, when I run `bundle install` with Bundler 2.2.11, it
remained stuck with the Intel-based version of the gem, unable to
understand it had to build from native extensions. Upgrading to the
latest version solved the issue. I decided not to investigate further.