Rename `Ios::L10nHelper` to `Ios::L10nLinterHelper`
As this helper is only containing code related to the SwiftGen-based linter and is already quite large, so we shouldn't add more to it.

I plan to put any other (upcoming) L10n-related helper methods (that I'll need for future actions around our L10n refactor project) in a separate helper, to avoid flooding this existing one with more stuff, hence renaming this one to a more ad-hoc name first in preparation.