Consolidate all `an_` and `gp_` update metadata specs in shared examples
Thanks for nudging in this direction @AliSoftware.

I had considered using shared examples but decided not to try it when I
noticed the difference with the `whats_new` block. Turns out that was
the only difference –that we found so far– and a simple flag was enough
to account for it in the shared examples

See
https://github.com/wordpress-mobile/release-toolkit/pull/326#discussion_r792947416