Fix support for unquoted .strings keys
Like in a typical InfoPlist.strings file

Also ensure those unquoted keys end up quoted after insertion of the prefix (which is what the `expected-merged-prefixed.strings` file already expected but was not really verifying because the source `InfoPlist-urf8-strings` file already had quotes before that commit). This is important in case adding the prefix would have made it otherwise invalid not to quote the key anymore (i.e. if the prefix use characters that are not valid as unquoted identifier)