Rename SiriIntentsTarget fixture to IntentsDefTarget
Since I learned recently that this .intentdefinition target that we have in WPiOS and I used as a model for this fixture... is not about Siri Intents after all, but only used for support of configuring the iOS 14 Widgets.

So I figured, better remove the mention of 'Siri' in the fixture as well to avoid confusion when trying to make sense of our tests and compare it with similar cases we have in real life.