Don't do the prefix post-processing if prefix = ''
The prefix post-processing relies on using RegExp and `gsub` on the lines of a `.strings` files in OpenStep/textual format.

Since it seems hard to find an official spec and formal grammar for that OpenStep format and that format can contain a lot of edge cases (quoted vs unquoted keys, values spanning multiple lines, character escapes, …), we are not fully confident that using a RegExp to try and parse those is fully trustworthy and will really covers all possible cases.

Which is why we want to limit this operation — of adding prefixes to keys in those textual-formatted `.strings` files — to the very minimum, and avoid this `gsub` post-processing whenever we can. Typically we will avoid adding prefixes to the keys from the main `Localizable.strings` file (which contain 99% of the keys) and only use a prefix for `InfoPlist.strings` or `Sites.strings` files (which usually have a very low number of keys).