Revert changes about imagemagick from pipeline.yml
Turns out `brew install imagemagick` is still necessary after all on CI, if not for running the tests themselves, at least for installing the gems from the Gemfile.

This is because:
 - Our `Gemfile` includes the `gemspec` directive, which includes all the gem dependencies declared in our `.gemspec` as part of the `Gemfile`.
   This is typical in a gem repo like here, as the goal is to ensure that working on and testing the codebase would always use the same dependencies as the gem itself uses.
 - Our `.gemspec` contains `rmagick` as one of its dependency
 - Which in turn needs the `imagemagick` formula/library to be installed (via `brew`), otherwise we will get an error during `bundle install`.

So in short, running `bundle install` will try to install all the gems from the `Gemfile`, which is a superset of the dependecies from the `.gemspec` and thus will try to install the `rmagick` gem, which requires `imagemagick` to be installed.