Address RuboCop violations
🤔 I wonder why my Vim setup didn't notify me about those locally

I also moved what I originally wanted to be the only public method at
the bottom of the file, thus restoring the original file structure. The
RuboCop warning made me realized the method to extract the key-lines
pairs, while currently unused, might be useful in the future. And since
this is but an internal helper, there's no harm in letting consumers see
it.