Use `in_tmp_dir` in place of `Dir.mktmpdir` in update metadata tests
This should have ensured that not only we create a temporary directory
but also `cd` into it to avoid scenarios where the state of the Git repo
affects the tests outcome (see the need for stubbing the Fastlane Git
helper in `ios_update_metadata_source_spec.rb`)

Unfortunately, this doesn't seem to do the trick. If the stub call is
removed, the test fails.