Fix `emulator` logs filtering
To still remove most of the noise, but still log ERROR and PANIC lines, especially to print the `PANIC: Broken AVD system path` and other boot errors like "cannot find kernel file" that ends up failing to boot the emulator.

If we don't print those. the user will see in their fastlane logs that the emulator command was called, then that the `adb wait-for-device` was called, but then adb will be waiting for that device forever since `emulator` never really launched.
In those cases it's thus critical to log the errors to help understand that it's no use to wait forever and to have hints of what to fix.