Use a `Thread` + `Open3.popen2` to launch emulator
So that we can process the stdout and stderr of the `emulator` process in Ruby, rather rhan relying on spawning a subshell and rely on piping to `sed` and having to use `system` and `&` for running in the background… which was not pretty.

Here by using `Thread.new` we're still able to launch the emulator in the background, but we're also able to process its `stdout` and `stderr` directly from our ruby code, by calling `Open3.popen2e`, reading the `stdout`+`stderr` as an `IO` line by line (`oe.readline`) as a line stream, and print emulator's log lines as they arrive to u to be able to filter then and add a prefix, all using Ruby code instead of `sed`

This also allows us to print the exit code of the `emulator` process if it didn't exit properly (non-zero status code).

As a final bonus, the `emulator` subprocess that has been opened by `Open3.popen2s` within that `Thread` will be automatically killed (being a child process) when the parent process (fastlane) terminates (successfully or not). Which was not the case when we used `system('emulator … &') to run the emulator as a background job.