Fix issue with retry mechanism
Making the previously-introduced test go green.

The issue was that by using `violations.merge` we retained the violations from each previous iterations and never cleaned them up in between loop iterations. So even if the user fixed the violation in the input file(s), the previous violations were still present in the `violations` hash with that old implementation.