Add ability to provide a direct download link
Useful for APK files for example.

PS: be mindful that providing a public link accessible without any access control (e.g. from a cloudfront public API or similar) means users outside Automattic might be able to download the APK even if they are not invited to App Center. Up to you to decide if that is ok or not when you choose to use this parameter.