Rename `update_branch_protection` to `set_branch_protection`
To account for the fact that if the branch is not protected, it will _set_ the default protection on the branch, while `update_branch_protection(branch: 'a_branch_that_was_not_protected')` didn't seem to make a lot of sense in that context.

If the branch is already protected, we can see the call as _setting_ the value of _individual_ branch protection settings/parameters (e.g. `lock_branch: true`), as opposed to _updating_ the overall branch protection. Which means `set_branch_protection` seems to cover more supported cases semantically than `update_branch_protection` does.