Use RSpec `double` instead of custom `FakeResponse`
Co-authored-by: Olivier Halligon <olivier.halligon@automattic.com>

Notice that this implementation is not ideal as far as RuboCop is
concerned.

The recommendation to use a verified double (`instance_double`) comes
with the intent of improving the test correctness by ensuring that the
expected object type is used in the system under test.

While that is an excellent general recommendation, in this particular
instance it backfires because the object `Buildkit` returns when
creating a build is `Sawyer::Resource` which we cannot as easily mock.

I.e., mocking with

```ruby
.and_return(
  instance_double(
    Sawyer::Resource,
    state: 'scheduled',
    web_url: expected_url
  )
)
```

Fails with:

```
the Sawyer::Resource class does not implement the instance method: state
```

I also tried to create a dummy `Sawyer::Resource` instance, but that
required providing a client object.

If this implementation is not satisfactory, we could consider stubbing
the Buildkite API HTTP response. I didn't go down that route when first
working on this because I didn't want to go look for endpoints and JSON
responses, but it's an option worth considering.