Remove deprecated `create_release` in favor of `create_github_release`
Build on top of
https://github.com/wordpress-mobile/release-toolkit/pull/585.

Those changes were merged just recently but their context was that of a
minor update, unaware that `trunk` already included a breaking change.

Since we are breaking things, I'd say we might as well remove backward
compatibility by forcing users that upgrade to 12.0.0 to rename
`create_release` to `create_github_release`.