Ensure no failure if none expected
Previously we only tested `expect(user_errors)` if `expected_failures` was not nil; but in practice even if `expected_failure` is `nil` we want to ensure no errors was raised.

Used the occasion to simplify the logic, given that we can always only expect at most one error to be raised (since raising an error exits the call early), so there was no need for this to be an array.

Also added some YARD doc on the helper to document the type and intent of its parameters