Remove period from "Update draft release notes..." commit message
The first line of a commit message should not have a period (`.`)

I am not aware of any official Git documentation for the statement
above. However, it has been a convention in all the projects and
companies I worked on.

The style is also present in the example commit message from the
official Git book. See
https://git-scm.com/docs/MyFirstContribution#:~:text=You%20will%20be%20presented%20with%20your,A%20U%20Thor%20%3Cauthor%40example.com%3E

The style is also used in other sources, such as:

- https://commit.style/
- https://www.gitkraken.com/learn/git/best-practices/git-commit-message#using-imperative-verb-form
- https://thoughtbot.com/blog/5-useful-tips-for-a-better-commit-message

One could argue that the 50 to 72 character length and the period
omission were characters economy constraints from back in the day of
small screens and text-only interfaces. While that is true, I find the
are many benefits in forcing oneself to be concise in the commit
title/summary. As for short line length, it works well for people like
me, who use big font sizes because their eyes strain otherwise.

But aside from all the justifications above, having commit message
titles without a period is the existing convention in this repo, which
make this change a no brainer.