Refactor `prototype_build_details_comment` action for Firebase App Distribution
App Center will be discontinued and we are migrating to Firebase App Distribution, so this action needs to derive its metadata from the `lane_context` exposed by the `firebase_app_distribution` action instead of the `appcenter_upload` one now.

Besides, less parameters are now needed, which allowed us to simplify the implementation and make it more readable.