Fix test so it passes on CI
When the tests are run from the Buildkite CI itself, the `BUILDKITE_PIPELINE_SLUG` env var would be set while the test runs, so the unit test case that is responsible for testing the case of `buildkite_pipeline_slug` parameter missing would not work as expected.

So we need to ensure we temporarily unset the env var for the duration of the test being run, so that it actually runs in a context we expect it to.